restructure project and move public files
