+++ js/lotto.js
